Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -4,7 +4,7 @@ struct Student
 {
     int roll;
     int name[50];
-    int marks;
+    float marks;
 };
 int main()
 {
